Move particle block into generators folder
