+++ packages/demo/src/configuration/blocks/generators/particleBlock.fragment.glsl
