+++ packages/demo/src/configuration/blocks/generators/particleBlock.ts
